// apps/amanda/src/utils/enums/log-level.ts
export enum LogLevel {
  /**
   * Represents the Error level.
   * Value: "ERROR"
   * Represents a priority of 3.
   */
  error = 'ERROR',
  /**
   * Represents the Warning level.
   * Value: "WARN"
   * Represents a priority of 2.
   */
  warning = 'WARN',
  /**
   * Represents the Debug level.
   * Value: "DEBUG"
   * Represents a priority of 1.
   */
  debug = 'DEBUG',
}